\name{circuit}
\alias{circuit}
\title{Initialize a circuit plot}
\description{Initialize a circuit plot}
\usage{circuit()}

%\arguments{}

\details{fill in more later}

\value{None.}

\seealso{Draw capacitors with \code{\link{capacitor}}.}

\examples{
library(circuit)
circuit()
capacitor(.5, .3, .5, horizontal=FALSE)
capacitor(.2, .2, .5, horizontal=TRUE)
}

\author{Dan Kelley}

\keyword{misc}
